+++ flake8_aaa/function.py
@@ -1,4 +1,5 @@
 from .act_block import ActBlock
+from .arrange_block import ArrangeBlock
 from .exceptions import NotActionBlock, ValidationError
 from .helpers import function_is_noop
 from .types import ActBlockType
@@ -11,12 +12,14 @@ class Function:
         node (ast.FunctionDef): AST for the test under lint.
     """
 
-    def __init__(self, node):
+    def __init__(self, node, file_lines):
         """
         Args:
             node (ast.FunctionDef)
+            file_lines (list (str)): Lines of file under test.
         """
         self.node = node
+        self.lines = file_lines[self.node.lineno - 1:self.node.last_token.end[0]]
         self.act_block = None
 
     def check_all(self):
@@ -33,6 +36,8 @@ class Function:
             return
 
         self.act_block = self.load_act_block()
+        self.arrange_block = self.load_arrange_block()
+        self.check_act_arrange_spacing()
 
     def load_act_block(self):
         """
@@ -61,3 +66,61 @@ class Function:
             raise ValidationError(self.node.lineno, self.node.col_offset, 'AAA02 multiple Act blocks found in test')
 
         return act_blocks[0]
+
+    def load_arrange_block(self):
+        """
+        Returns:
+            ArrangeBlock: Or ``None`` if no Act block is found.
+
+        Raises:
+            ValidationError
+        """
+        arrange_block = ArrangeBlock()
+        for node in self.node.body:
+            if node == self.act_block.node:
+                break
+            arrange_block.add_node(node)
+
+        if len(arrange_block.nodes) > 0:
+            return arrange_block
+
+        return None
+
+    def check_act_arrange_spacing(self):
+        """
+        When Function has an Arrange block, then ensure that there is a blank
+        line between that and the Act block.
+
+        Returns:
+            None
+
+        Raises:
+            ValidationError: When no space found.
+
+        Note:
+            Due to Flake8's error ``E303``, we do not have to check that there
+            is more than one space.
+        """
+        if self.arrange_block:
+            line_before_act = self.get_line_relative_to_node(self.act_block.node, -1)
+            if line_before_act != '\n':
+                raise ValidationError(
+                    line_number=self.act_block.node.lineno,
+                    offset=self.act_block.node.col_offset,
+                    text='AAA03 expected 1 blank line before Act block, found none',
+                )
+
+    def get_line_relative_to_node(self, target_node, offset):
+        """
+        Args:
+            target_node (ast.node)
+            offset (int)
+
+        Returns:
+            str
+
+        Raises:
+            IndexError: when ``offset`` takes the request out of bounds of this
+                Function's lines.
+        """
+        return self.lines[target_node.lineno - self.node.lineno + offset]