+++ tests/function/test_init.py
@@ -4,11 +4,21 @@ from flake8_aaa.function import Function
 
 
 @pytest.mark.parametrize('code_str', ['''
+
 def test():
     pass
+
+def test_other():
+    pass
+
+# All done :D
 '''])
-def test(first_node_with_tokens):
-    result = Function(first_node_with_tokens)
+def test(first_node_with_tokens, lines):
+    result = Function(first_node_with_tokens, lines)
 
     assert result.node == first_node_with_tokens
+    assert result.lines == [
+        'def test():\n',
+        '    pass\n',
+    ]
     assert result.act_block is None